feat: refactor test step button
